+++ CHANGES.rst
@@ -54,6 +54,7 @@ Bugfixes
   (:pr:`6730`)
 - Fix the usage of select list scrollbar causing it to close immediately (:issue:`6735`,
   :pr:`6736`, thanks :user:`foxbunny`)
+- Trigger event creation notification emails when cloning events (:pr:`6744`)
 
 Accessibility
 ^^^^^^^^^^^^^